Every PHP session is now stored on REDIS for redundancy
All session are stored in REDIS a centralized session storage. The fact that all user sessions are stored in an external container other the UI_SVC provides fail-over redundancy. If a ui_svc instance fails then the session will be transferred together with connection to the new ui_svc instance specified by the HAProxy