Worker task assignment now happens async
When a worker is requested to handle a task, the execution flow of the task happens in an async way using then CompletableFuture java class. That means that the worker will reassure the monitor that its gonna handle the task but the acknowledgement wont wait for the completion of the task. Worker now is technically multithreaded